+++ .gitignore
@@ -1,9 +1,10 @@
-/node_modules/*
+/.vscode/*
+node_modules
 package-lock.json
 /data/exported/*.class
 /data/exported/*.java
 /data/scanned/*.jpg
 /data/scanned/*.png
-/src/Backend/*.class
-/src/User/*
-!/src/User/placeholder.txt
+/src/backend/Java/*.class
+/src/backend/Java/user/*
+!/src/backend/Java/user/placeholder.txt
