+++ src/frontend/public/favicon.ico
