+++ src/frontend/public/logo192.png
