+++ src/frontend/public/logo512.png
