+++ src/frontend/public/manifest.json
