+++ start_webserver.bat
@@ -1,2 +1,3 @@
-node .
+cd ./src
+node server.js
 pause
